requirements.lock.txt
absl-py==2.3.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
farama-notifications==0.0.4
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
grpcio==1.73.1
gymnasium==1.1.1
jinja2==3.1.6
kiwisolver==1.4.8
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyboy @ git+https://github.com/Baekalfen/PyBoy.git@02fe4f3b363bae413d9b81f34338c110c49ff409
pyparsing==3.2.3
pysdl2==0.9.17
pysdl2-dll==2.32.0
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
stable-baselines3==2.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.7.1
triton==3.3.1
typing-extensions==4.14.1
tzdata==2025.2
werkzeug==3.1.3
